multiple docstring updates
